+++ src/sip_parser.h
@@ -20,6 +20,6 @@
 
 int sip_message_parser(const char *incoming_sip_msg, size_t packet_size,
 		       bad_actor *bad_actor_event,
-		       struct sentrypeer_config const *config);
+		       sentrypeer_config const *config);
 
 #endif //SENTRYPEER_SIP_PARSER_H