+++ tests/unit_tests/test_conf.h
@@ -1,6 +1,6 @@
 /* SPDX-License-Identifier: GPL-2.0-only OR GPL-3.0-only  */
 /* Copyright (c) 2021 Gavin Henry <ghenry@sentrypeer.org> */
-/*
+/* 
    _____            _              _____
   / ____|          | |            |  __ \
  | (___   ___ _ __ | |_ _ __ _   _| |__) |__  ___ _ __
@@ -11,13 +11,9 @@
                              |___/
 */
 
-#ifndef SENTRYPEER_HTTP_DAEMON_H
-#define SENTRYPEER_HTTP_DAEMON_H 1
+#ifndef SENTRYPEER_TEST_CONF_H
+#define SENTRYPEER_TEST_CONF_H 1
 
-#include "conf.h"
+void test_conf(void **state);
 
-#define HTTP_DAEMON_PORT 8082
-
-int http_daemon_init(sentrypeer_config const *config);
-
-#endif //SENTRYPEER_HTTP_DAEMON_H
+#endif //SENTRYPEER_TEST_CONF_H